# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/sse": {
      "get": {
        "operationId": "ServerEventsController_sse",
        "parameters": [],
        "responses": {
          "200": {
            "description": "SSE stream"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Establish SSE connection for server events",
        "tags": [
          "SSE"
        ]
      }
    },
    "/serial": {
      "get": {
        "description": "Retrieve a list of serial port info objects",
        "operationId": "SerialController_getList",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Port info list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "Returns serial ports info",
        "tags": [
          "Serial"
        ]
      },
      "delete": {
        "description": "Delete all serial ports and associated universes",
        "operationId": "SerialController_dropSerial",
        "parameters": [],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "Deletes all serial ports and universes",
        "tags": [
          "Serial"
        ]
      },
      "post": {
        "description": "Create a new serial port and associated universe",
        "operationId": "SerialController_addSerial",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Serial options",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "driver": {
                    "type": "string",
                    "description": "The serial driver",
                    "example": "serial"
                  },
                  "port": {
                    "type": "string",
                    "description": "The serial port",
                    "example": "/dev/ttyUSB0"
                  },
                  "baudrate": {
                    "type": "number",
                    "description": "The baud rate",
                    "example": 115200
                  },
                  "universe": {
                    "type": "number",
                    "description": "The universe",
                    "example": 1
                  }
                },
                "$ref": "#/components/schemas/"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "summary": "Adds a new serial port and associated universe",
        "tags": [
          "Serial"
        ]
      }
    },
    "/serial/{id}": {
      "delete": {
        "description": "Delete a serial port and associated universe by ID",
        "operationId": "SerialController_deleteSerial",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The ID of the serial port",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "summary": "Deletes a serial port and its associated universe",
        "tags": [
          "Serial"
        ]
      }
    },
    "/serial/current": {
      "get": {
        "description": "Retrieve the ID of the current serial port",
        "operationId": "SerialController_getCurrent",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current serial port ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "summary": "Returns the ID of the current serial port",
        "tags": [
          "Serial"
        ]
      },
      "post": {
        "description": "Set the current serial port by ID",
        "operationId": "SerialController_setCurrent",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "The ID of the serial port to set as current",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "summary": "Set Current Serial Port",
        "tags": [
          "Serial"
        ]
      }
    },
    "/serial/drivers": {
      "get": {
        "description": "Retrieve a list of supported serial drivers",
        "operationId": "SerialController_getDrivers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Driver list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array"
                }
              }
            }
          }
        },
        "summary": "Returns supported driver list",
        "tags": [
          "Serial"
        ]
      }
    },
    "/dmx": {
      "get": {
        "operationId": "DMXController_getUniverses",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Universes id list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "Returns universes list",
        "tags": [
          "DMX"
        ]
      },
      "delete": {
        "operationId": "DMXController_deleteAllUniverses",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete all universes",
        "tags": [
          "DMX"
        ]
      },
      "post": {
        "operationId": "DMXController_addUniverse",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Universe options",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Add universe",
        "tags": [
          "DMX"
        ]
      }
    },
    "/dmx/{id}": {
      "delete": {
        "operationId": "DMXController_deleteUniverse",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Universe id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete universe",
        "tags": [
          "DMX"
        ]
      }
    },
    "/dmx/{id}/values": {
      "get": {
        "operationId": "DMXController_getValues",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Universe id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of DMX values",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array"
                }
              }
            }
          }
        },
        "summary": "Get DMX values",
        "tags": [
          "DMX"
        ]
      }
    },
    "/dmx/{id}/values/{value}": {
      "patch": {
        "operationId": "DMXController_setValues",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Universe id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "value",
            "required": true,
            "in": "path",
            "description": "New DMX channels value",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Set DMX values",
        "tags": [
          "DMX"
        ]
      }
    },
    "/dmx/{id}/channel": {
      "patch": {
        "operationId": "DMXController_setChannel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Universe id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "New DMX values",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Sets several DMX channels",
        "tags": [
          "DMX"
        ]
      }
    },
    "/dmx/{id}/channel/{address}": {
      "get": {
        "operationId": "DMXController_getChannelValue",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Universe id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "address",
            "required": true,
            "in": "path",
            "description": "DMX channel address",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "DMX channel value",
            "content": {
              "application/json": {
                "schema": {
                  "type": "number"
                }
              }
            }
          }
        },
        "summary": "Get DMX channel value",
        "tags": [
          "DMX"
        ]
      }
    },
    "/dmx/{id}/channel/{address}/value/{value}": {
      "patch": {
        "operationId": "DMXController_setChannelValue",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Universe id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "address",
            "required": true,
            "in": "path",
            "description": "DMX channel address",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "value",
            "required": true,
            "in": "path",
            "description": "DMX channel value",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Set DMX channel value",
        "tags": [
          "DMX"
        ]
      }
    },
    "/device": {
      "get": {
        "operationId": "DeviceController_getDevices",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Map {index: device}",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "Returns device list",
        "tags": [
          "Device"
        ]
      },
      "delete": {
        "operationId": "DeviceController_removeAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete all devices",
        "tags": [
          "Device"
        ]
      },
      "post": {
        "operationId": "DeviceController_add",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Device object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "DMX address of device",
            "content": {
              "application/json": {
                "schema": {
                  "type": "number"
                }
              }
            }
          }
        },
        "summary": "Add new device",
        "tags": [
          "Device"
        ]
      },
      "patch": {
        "operationId": "DeviceController_set",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Device list",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Set bulk devices",
        "tags": [
          "Device"
        ]
      }
    },
    "/device/{index}": {
      "get": {
        "operationId": "DeviceController_get",
        "parameters": [
          {
            "name": "index",
            "required": true,
            "in": "path",
            "description": "Device index",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Device object",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "Get device",
        "tags": [
          "Device"
        ]
      },
      "patch": {
        "operationId": "DeviceController_update",
        "parameters": [
          {
            "name": "index",
            "required": true,
            "in": "path",
            "description": "Device index",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Device object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update device",
        "tags": [
          "Device"
        ]
      },
      "delete": {
        "operationId": "DeviceController_remove",
        "parameters": [
          {
            "name": "index",
            "required": true,
            "in": "path",
            "description": "Device index",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete device",
        "tags": [
          "Device"
        ]
      }
    },
    "/device/dmx/{index}": {
      "get": {
        "operationId": "DeviceController_getDMXChannel",
        "parameters": [
          {
            "name": "index",
            "required": true,
            "in": "path",
            "description": "Device index",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get DMX channel",
        "tags": [
          "Device"
        ]
      }
    }
  },
  "info": {
    "title": "@dmx-cloud/dmx-backend",
    "description": "",
    "version": "0.0.1",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "Object": {
        "type": "object",
        "properties": {}
      }
    }
  }
}